Core/PacketIO: Updated to 11.0.0
Port From (https://github.com/TrinityCore/TrinityCore/commit/e59eef5432c7b70679d33f4911c88d0f7d75fd39)

+++ Source/Game/Entities/Item/Item.cs
@@ -2342,7 +2342,7 @@ namespace Game.Entities
         }
 
         public ItemContext GetContext() { return (ItemContext)(int)m_itemData.Context; }
-        public void SetContext(ItemContext context) { SetUpdateFieldValue(m_values.ModifyValue(m_itemData).ModifyValue(m_itemData.Context), (int)context); }
+        public void SetContext(ItemContext context) { SetUpdateFieldValue(m_values.ModifyValue(m_itemData).ModifyValue(m_itemData.Context), (byte)context); }
 
         public void SetPetitionId(uint petitionId)
         {
